fix: 코스데이터 등록시 버튼 비활성화 (#114) - 코스등록시 네트워크 연결상태에 따라서 등록버튼 눌림 방지 - 거리계산 단위 통합 진행중

+++ TrackUs/TrackUs/Sources/Running/View/RunningResultView.swift
@@ -25,7 +25,7 @@ struct RunningResultView: View {
     }
     
     // 예상시간
-    var estimatedTime: Int {
+    var estimatedTime: Double {
         return ExerciseManager.calculateEstimatedTime(distance: targetDistance * 60)
     }
